The Api security is strength and new Pages added UI part

+++ ClassTrack/Presentation/ClassTrack.MVC/Controllers/ClassController.cs
@@ -8,10 +8,10 @@ namespace ClassTrack.MVC.Controllers
 {
     public class ClassController : Controller
     {
-        private readonly IClassRoomService _roomService;
+        private readonly IClassRoomClientService _roomService;
         private readonly IAuthenticationClientService _clientService;
 
-        public ClassController(IClassRoomService roomService,
+        public ClassController(IClassRoomClientService roomService,
                                 IAuthenticationClientService clientService)
         {
             _roomService = roomService;
@@ -26,8 +26,16 @@ namespace ClassTrack.MVC.Controllers
            return RedirectToAction("Index","Home");
         }
         public async Task<IActionResult> ClassRoom(long id)
-        {          
-            return View(await _roomService.GetByIdAsync(id));
+        {
+            if (id < 1)
+                return BadRequest();
+
+            GetClassRoomWithPermissionVM getClassRoom = await _roomService.GetByIdAsync(id);
+
+            if (getClassRoom.IsTeacher.IsTeacher)
+                return View("TeacherClassRoom", getClassRoom);
+
+            return View("StudentClassRoom",getClassRoom);
         }
 
         public async Task<IActionResult> Dashboard()
@@ -44,20 +52,29 @@ namespace ClassTrack.MVC.Controllers
                                                     dashboardVM.PostClass));
             }
 
+            if(dashboardVM.PostClass.Name.Length > 150)
+            {
+                ModelState.AddModelError(nameof(PostClassRoomVM.Name),
+                                        "The Name Length is too Long!");
+                return View(new DashboardVM(await _roomService.GetAllAsync(),
+                                                  dashboardVM.PostClass));
+            }
+              
             await _roomService.CreateClassRoomAsync(dashboardVM.PostClass);
-
             return RedirectToAction("Dashboard");
         }
 
        
-        [HttpPost]
+       
         public async Task<IActionResult> DeleteClassRoom(long id)
         {
+            if (id < 1)
+                return BadRequest();
+
             await _roomService.DeleteClassRoomAsync(id);
             return RedirectToAction("Dashboard");
         }
-
-        
+   
 
         public IActionResult QuizEditor()
         {
